requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
configparser==5.0.0
fxcmpy==1.2.6
idna==2.9
numpy==1.18.3
pandas==1.0.3
PySocks==1.7.1
python-dateutil==2.8.1
python-engineio==3.12.1
python-socketio==4.5.1
pytz==2020.1
requests==2.23.0
six==1.14.0
socketIO-client==0.7.2
urllib3==1.25.9
websocket-client==0.57.0
